feat: Change HCL to YAML
HCL is a powerful language but brings a lot of complexity and
interoperability issues between languages and stacks.

CloudQuery doesn't use any of the powerful HCL features and mainly uses
it for very simple configs.

Moreover YAML and Jsons have good ecosystems of schema validators which will
remove the need for code validation and will also help streamline providers configuration.