feat: Add support for Tenants (#412)
As requested in https://github.com/cloudquery/cq-provider-azure/issues/406, this adds support for listing tenants under a subscription. 

It can be used by adding `subscriptions.tenant` to the cloudquery config, which will then populate a table named `azure_subscription_tenants`.

To add this, we had to start using the new `armsubscriptions` Azure SDK, which requires Go 1.18 and uses generics. The implementation for the subscriptions client has been switched to use this now. The benefit of this is that we also get new columns and relations on the `accounts` and `subscriptions` tables. 

Tables updated in this PR:
 - `azure_account_location_paired_region` (new)
 - `azure_account_locations` (new columns)
 - `azure_subscription_subscriptions` (new columns, notably `tenant_id` and `managed_by_tenants`)
 - `azure_subscription_tenants` (new)

Closes https://github.com/cloudquery/cq-provider-azure/issues/406