fix(aws): Fix cloudtrail status column not defined (#1892)
This should fix undefined status field in cloudtrail. in the future we
can also unnest it but given it seems it was never defined and was
silently ignored we can do it like this for now (though need to go back
in the old code). In the future we can add support for code gen to
generate tables from multiple structs.

Closes https://github.com/cloudquery/cloudquery/issues/1882

Co-authored-by: Herman Schaaf <hermanschaaf@gmail.com>