feat(dest-test): Add new test destination (#2057)
Adding a new test destination for CLI testing purposes. The new
destination plugin is basically a dev null.

Also, Im proposing/trying out a .yml per plugin ? maybe this way we can
take advantage of the built-in github

```
    paths:
      - "plugins/destination/test/**"
```

For destination plugins I think this is specifaclly important because
the workflows vary very much between each plugin as we need different
databases and so on.

I think for source plugin this will also start becoming a thing when we
introduce some smoke tests that will require credentials and so on.