chore: Make GH workflow per plugin/cli
This is a take at a different approach where we use workflow
per plugin. This should give the following advantages:

- native workflow filtering: actions wont run if files not changed
which will save lot's of time, GH minutes and money
- we still have the ability to easily customize workflow per plugin
when needed. especially true for destinations but also for sources.
- if we have any common functionatliy we can always put it in scripts
and then reuse them in the actions rather then colliding the actions.