chore(cli): Make sentry only send errors on panic
Go doesn't really have a great way of distinguishing exceptions
and errors. Also we don't want to go into the rabbit hole of
classifying every error as user error or internal error as sometimes
it's not really possible: Let's say we had an error in our
marshaling logic where usually this error means a user error in some cases
it can be also an internal one.

We will have to rely on user reports for the most, apart from panics
which are clearly internal errors and we would like to send them to sentry.