fix(cli): Close stdout when using --log-console (#2459)
when we use --log-console we dont want standard output. logs are by default written to stderr so we just close stdout which make any fmt.Print* silent. This makes it easy to use fmt.Print* in the CLI and not use any new abstractions and prop drilling.

This also adds handling of ctrl+c, that should go together with this as well https://github.com/cloudquery/plugin-sdk/pull/252