chore: Add step to update versions in docs (#2477)
This adds steps to the plugin release workflow to update the versions in the docs to the latest. It currently handles two cases:
 - a string like `version: "v2.0.0" # latest version of aws plugin` will get matched and the version replaced with the latest version.
 - a string in the sources/destinations table will get matched and replaced with the latest version for easy reference.

I tried some other, less janky, approaches but couldn't get it working with nextra within a couple of hours and decided to revert to a simpler solution.