feat: Add support for EC2 reserved instances (#1912)
This adds support for fetching EC2 reserved instances. 

It is meant to address https://github.com/cloudquery/cloudquery/issues/1815, but please note that since this change is being applied to (the as-of-yet unreleased) CloudQuery v2, it will require upgrading to the new version once it is released.

I couldn't see a way to paginate through these results, so I hope that the API returns the full resultset for each region in the first call.