fix: Add codegen for AWS regions (#2701)
Found another case of an AWS resource not being generated with codegen. It was unfortunately not so easy to deal with this one in a backwards-compatible way. The biggest issue is keeping the name `aws_regions` rather than `aws_ec2_regions` (which would be in line with all other resources). 

Note: This could perhaps be considered a `chore`, as it should change no user-facing functionality, but since there is a small change in the ordering of the columns I think I'd rather have it reported as one of the changes in the next release.