chore(docs): AWS: Add missing configuration and example config (#3056)
This adds an example config to the top of the AWS config page to better place it in context. It also adds two options under `accounts` that were undocumented before.

Also, reverting the `Organization -> org` change as I think that was misread, the config option is called `org`, but the object is an `Organization` object.