chore(aws): Autogenerate service interfaces and other common structures (#3494)
This removes several error-prone steps for contributing new AWS resources and replaces them with code generation. 

## Before
 - service interfaces had to be created manually by copy-pasting signatures from the AWS SDK
 - the `go:generate mockgen` comment had to be added manually
 - `go generate` had to be run separately to generate mocks
 - the service had to be added in `client/client.go`
 - the service had to be instantiated in `initServices`
 - the table had to be registered in `resources/plugin/tables.go`

## Now
 - a Client struct can be added to `codegen/services/clients.go`
 - `make gen` does all of the previously manual steps for you

## Notes for reviewers

This PR touches a lot of files, because I decided it would be better to generate consistent names than maintain a list of names just for Go-style capitalization. It's almost entirely renaming of variables, there are no functional changes.