fix!: Make path a required config parameter (#3872)
This SDK upgrade to the CLI adds the code in https://github.com/cloudquery/plugin-sdk/pull/371, and as such will make `path` a required config parameter. We consider this a breaking change, so the CLI version will be bumped to 2.0.0. 

As context, we are making `path` required (and not inferred from `name`) because there has been some confusion around these settings. Previously `path` could have been inferred from `name` in some circumstances, but when using multiple configs, this was not always possible, causing confusion.

From now on, `path` is a required parameter. It will usually take the form of `cloudquery/aws`, where `aws` is the source plugin name in this example. For the postgresql destination, it will be `cloudquery/postgresql`.

- `name` should be thought of as the unique identifier of the particular source/destination config. It needs to be unique, as it will be used for overwriting stale entries in `overwrite-delete-stale` mode, inserted as a column alongside all entries, and ensure there are no race conditions when running multiple configs in parallel.
- `path` is normally the unique path to download the plugin from. We have some special conventions around this, documented [here](https://www.cloudquery.io/docs/developers/creating-new-plugin#naming-conventions). It can also be used during development to point to a locally-running binary or GRPC server.

Related:
 - https://github.com/cloudquery/plugin-sdk/issues/317
 - https://github.com/cloudquery/plugin-sdk/issues/256