chore(codegen): Replace CloudQueryServiceName with function (#3904)
This removes one parameter from codegen for resolvers and mocks. 

Service name should now be consistent enough that capitalizing the first letter should work in 99% of cases. The service name should almost always be the same as the aws package name, which is where CloudQuery service name is derived from.