feat(aws): Debug logging to include body (#3964)
This was required to debug the #3939 issue, might be useful for others... Could also consider introducing a `aws_debug_mode` integer, but AWS SDK's log flags are high 64-bit bitfields so the ints (if exposed as-is) look a bit too high.

```go
// Supported ClientLogMode bits that can be configured to toggle logging of specific SDK events.
const (
	LogSigning ClientLogMode = 1 << (64 - 1 - iota)
	LogRetries
	LogRequest
	LogRequestWithBody
	LogResponse
	LogResponseWithBody
	LogDeprecatedUsage
	LogRequestEventMessage
	LogResponseEventMessage
)
```

ie. LogSigning resolves to -9223372036854775808, LogRetries to 4611686018427387904, and so on.