feat: Regenerate k8s (#4495)
Given multiple requests from number of key users around incompleteness  of k8s I updated our codegen to do both fetcher and mocks. k8s API is especially good and consistent so it was relatively straight forward to have 99% auto-generated source plugin apart of a few places with custom resolver or overrides (that are still done on the codegen level).

This doesn't have any breaking changes but things to consider in the future is how to handle v1,v1beta,v2 apis. We might want to introduce those into our tables as well but can be done in a future PR.

Once this is merged it will also help with the WIP policy for k8s.