fix: Remove codegen files before re-generating for GCP (#4720)
Similar to https://github.com/cloudquery/cloudquery/pull/4719, this adds a step to the gen-code makefile command for GCP to remove all generated files first. This helps avoid issues where generated files are left in the repo despite no longer being part of any recipes. 

There was one such issue which is fixed here. The mock file is different when generated, and it seems like manual edits have been made to it, so I've removed the `// Code generated by codegen; DO NOT EDIT.` comment from the top of the file.