fix: Allow sync to complete when one or more accounts fail to authenticate (#5030)
Closes https://github.com/cloudquery/cloudquery/issues/5028

I tested this locally using the config shown in the original ticket. If
all accounts fail to auth, the CLI returns with exit code 1. If one
account succeeds, an error is logged for the failing accounts. The error
is not counted as part of the `Errors` count in the output, but in any
case that would likely be lost in the count of resolver-related errors
and quite hard to implement right now. Users will need to discover such
errors through the `ERR` line in the logs, which I think is not
unreasonable.