fix: Add refresh window and jitter for cached credentials (#5097)
This is meant to address https://github.com/cloudquery/cloudquery/issues/4964. It enables the credentials cache to refresh its credentials with the underlying provider before the hard timeout is reached.

The exact error reported in the issue was:

```
failed to sign request: failed to retrieve credentials: failed to refresh cached credentials, operation error STS: AssumeRole, https response error StatusCode: 403, RequestID: 4415c3fc-eb88-4b2d-b955-1dbe40693f78, api error ExpiredToken: The security token included in the request is expired
```

With emphasis on the "failed to refresh cached credentials" part and the **ExpiredToken** error, this should address a potential race condition with expiring credentials that can cause the refresh request to fail unexpectedly due to ExpiredToken exceptions.

The underlying issue in the original ticket could be something else (I have so far been unable to reproduce it), but this change seems like a good thing to include in any case.