fix: Filter results of getTablesForSpec to top-level tables in CLI (#5294)
This fixes a hanging bug in recent versions of the CLI and some destination plugins (e.g. BigQuery, Snowflake). The actual bug is in the `GetTablesForSpec` SDK function, which should not return a flattened list of tables, but this fix for the CLI means the CLI will work with any version of `GetTablesForSpec` (even versions with the bug). 

I will follow up shortly with a fix for the SDK.