chore: Add policy documentation to website (#5296)
This adds a script to extract from policies:
 1. the tables used
 2. the query titles
and then write these to a `Policies` page on the website:
![Screenshot 2022-12-02 at 17 04 46](https://user-images.githubusercontent.com/1121616/205346438-b5c63bd6-6225-43e6-bbbf-d9cf6e81ea1d.png)

I also ended up updating a number of policies. In some cases there were missing titles, and in some cases there were queries that had syntax errors and were not being used by any policy. I also added `as title` to many queries (and I think this aids readability), but ultimately realized there are too many instances of queries without it, so made the script handle cases where `as title` is not specified.