fix: Fix deadlock in BigQuery destination if error occurs during write (#5550)
This fixes a deadlock that occurs if the BigQuery destination encounters an error that causes one of the writers to fail. Previously, the loop would have continued to try and send items to the channel, but these would not get processed. We now check that `gctx` is not done, and exit the loop either when `gctx` is done (so a worker error occurred) or `r, ok := <-res` has `ok` as false, which means the `res` channel was closed.

This issue might also happen in some other destinations, I will take a look there shortly.