feat(codegen): Add PrimaryKeys field to codegen (#5623)
#### Summary

Looks like we're using `ExtraColumns` mostly for setting primary keys, so I changed it so we only pass the names of primary keys. This allowed to me to remove some boilerplate around skipping the field, just so we can extra column it.
Also this saves copy pasting the column type, resolver (which was missing in some cases), etc.

**While the order of primary keys columns changed between all columns, the order between primary keys of the same table is the same**.

I'm not sure how to classify this? `feat`? It is user facing, both in regards to columns and code gen.

<!--