feat(postgresql): Update pgx to v5 (#5757)
This is a follow-up to https://github.com/cloudquery/cloudquery/pull/5733 (see discussion there). and needed so we can incorporate this nice bit: https://github.com/jackc/pgx/pull/1441

The https://github.com/jackc/pgx/pull/1441 is not a pre-requirement for this to go through. even better we should prob ship this first and see that `v5` works for us and for our users and if not we can do the same fix for `v4` (we will just have to fork it as I don't think `v4` accepts anymore contributions)