feat(salesforce): New Salesforce source plugin (#5954)
This is a new salesforce source plugin.

Right now this plugin uses one table and stays true to the API with a table that have an `id` (pk), `type` (pk), and `raw` (raw response from rest endpoint).

It is possible to turn it into multi-table but it might involve additional change in the SDK to allow dynamic tables as this is how the API behaves. If this will be need we can then "unwrap" the columns in each table.