fix(aws): Remove account validation (#6226)
#### Summary

Leaving this open for tomorrow. A user on Discord reported that a previously working configuration stopped working on `v9.0.1`.

Relevant configuration:

```yaml
  kind: source
  spec:
    name: aws
    path: cloudquery/aws
    version: "v9.0.1"
    destinations: ["postgresql"]
    tables: ["*"]
    skip_tables:
** snipped for copy/paste **
    spec:
      regions:
        - us-east-1
        - us-east-2
        - us-west-1
        - us-west-2
      accounts:
        - id: "services"
          role_arn: "arn:aws:iam::XXXXXXXXXXXX:role/security/cloudquery-ro-role"
          role_session_name: "cloudquery_inventory"
        - id: "prod"
          role_arn: "arn:aws:iam::XXXXXXXXXXXX:role/security/cloudquery-ro-role"
          role_session_name: "cloudquery_inventory"
```

Error:

```
spec validation failed: invalid accounts: invalid account id: services (should be 12 digits)"
```

It seems users should be able to use the alias of an account in the configuration. Keeping open for discussion and so that users can subscribe to get updates on this.
Also not sure about the `skip_member_accounts` validation (I kept that one). 

<!--