feat: Move azure to avoid codegen. (#6276)
This shouldn't introduce any public facing changes.

This should go together with:
https://github.com/cloudquery/plugin-sdk/pull/564

In general codegen creates a sub-optimal developer experience where you
have to juggle between `codegen` and code to make any code-changes. Not
only it creates a tough experience for the developer it also makes it
hard to review and to contribute.

This take here moves codegen only to the discovery phase and moves the
transformation to be part of the SDK which will eventually deprecate the
`codegen` package from the SDK completely and will avoid having two APIs
which are almost identical but quite the same.