fix(bigquery): Migrations: ignore extra columns in the table (#6370)
BigQuery migrations should not fail if there are extra columns present in the table. As long as the columns are not required, it should ignore those columns and add any extra ones required by CloudQuery table schemas. This allows users to add additional columns of their own, and it means that you can upgrade to a new minor version of a source plugin and roll back later without issues.

This is tested in the new tests added in https://github.com/cloudquery/plugin-sdk/pull/574

- Fixes https://github.com/cloudquery/cloudquery/issues/6364