fix: Fix SQLite migration logic (#6372)
There was a bug in the query that checks whether a table exists - it always found the table to not exist because `?` was being quoted. We also now select only the columns defined in the table schema, rather than all columns.

This will allow https://github.com/cloudquery/plugin-sdk/pull/574/files to pass.