feat: Move AWS to use built-in transformations (no codegen) (#6337)
Similar to https://github.com/cloudquery/cloudquery/pull/6276 moving
transformation to be part of our SDK
https://github.com/cloudquery/plugin-sdk/pull/564

This is related to https://github.com/cloudquery/plugin-sdk/pull/573 

Few notes on column changes:

most of changes are just additional columns and this should be fine as
we did too much transformation. we want to keep the code simple, focus
on extraction and provide and easy way of doing "dumb" one level
transformation via reflection. The additional column are because we have
in some places a duplicate where we add our own `arn` so it can be
consistent but this is our decision and we still want to be as close as
possible to the API so some duplicates is fine (and it makes the code
simple and consistent).

breaking changes in `aws_docdb_pending_maintenance_actions ` those
apparently never worked and those columns where always empty (they
didn't have mock tests so we didn't catch those)

Co-authored-by: Kemal <223029+disq@users.noreply.github.com>
Co-authored-by: Kemal Hadimli <disq@users.noreply.github.com>