doc(website): Adding how-to guide for using Renovate (#6435)
#### Summary

Adding a `how-to` guide for leveraging [Renovate](https://docs.renovatebot.com/) to keep CloudQuery plugins up-to-date per https://github.com/cloudquery/cloudquery/issues/6335. I did follow the setup instructions and ran the website locally and everything appears to load fine on my end the only exception being the header which renders a weird space:

![image](https://user-images.githubusercontent.com/95309754/211044519-a9f7e0a5-bef0-41f2-b9ca-056e7605b026.png)

but that's due to the `team.ts` content being loaded that populates that with the avatar/links for CloudQuery employees.