fix(gcp-project-policies): Use correct API to get Policy v3, fix policy 2.1 query (#7053)
<!-- 🎉 Thank you for making CloudQuery awesome by submitting a PR 🎉 -->

#### Summary

Reported on Discord
https://discord.com/channels/872925471417962546/1065556442800734268/1065871250225909791

`projectsClient.GetIamPolicy` returns `v1.Policy` which has `log_type`
as an enum, and as a result evaluated to an int in the destination
database. Our query needs the strings, as they appear in `v3.Policy`
(configured by the table).
To get the `v3.Policy` we need the protobuf API.

Additionally the check was wrong. We should verify all logging types are
enabled for `allServices`.
This is how it looks in GCP:

![image](https://user-images.githubusercontent.com/26760571/214112370-0badc1b4-4465-4b7e-973b-7f08f2a7fa57.png)

<!--
Use the following steps to ensure your PR is ready to be reviewed

- [ ] Read the [contribution guidelines](../blob/main/CONTRIBUTING.md)
🧑‍🎓
- [ ] Test locally on your own infrastructure
- [ ] Run `go fmt` to format your code 🖊
- [ ] Lint your changes via `golangci-lint run` 🚨 (install golangci-lint
[here](https://golangci-lint.run/usage/install/#local-installation))
- [ ] Update or add tests 🧪
- [ ] Ensure the status checks below are successful ✅
--->

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>