feat: Support Kinesis Firehose as a Destination (#6954)
#### Summary

Supports Kinesis Firehose as a destination...

Code is working, but not sure how to test as Kinesis Firehose Doesn't have a capability to read from it. The closest we could do is set batching to the lowest setting and wait 5 minutes for data to appear in the destination. But there is no way to control for multiple tests running at the same time.


<!--