chore(ci): Add workflow to set sentry alerts to Slack (#7617)
<!-- 🎉 Thank you for making CloudQuery awesome by submitting a PR 🎉 -->

#### Summary

In order to send Sentry alerts to Slack we need to create one per
project (i.e. plugin). That's a bit tedious so this workflow does it for
us periodically so new plugins won't be missing alerts.
If the alert exists it skips it (so it won't override any manual
configuration).

<!--
Use the following steps to ensure your PR is ready to be reviewed

- [ ] Read the [contribution guidelines](../blob/main/CONTRIBUTING.md)
🧑‍🎓
- [ ] Test locally on your own infrastructure
- [ ] Run `go fmt` to format your code 🖊
- [ ] Lint your changes via `golangci-lint run` 🚨 (install golangci-lint
[here](https://golangci-lint.run/usage/install/#local-installation))
- [ ] Update or add tests 🧪
- [ ] Ensure the status checks below are successful ✅
--->