fix: Fix ECR image ARN to use repository name (#7839)
The ECR image ARN was using `RegistryId`, which actually maps to the account ID. The repository ARN uses the repository name. I don't think repository images really have ARNs, but this change should allow this column to be used as a PK for now without the possibility of the same image in different repositories causing rows to be overwritten.