chore(website): Update integration logos (#8165)
Before:

![Screenshot 2023-02-16 at 16 51 33](https://user-images.githubusercontent.com/1121616/219433224-30855eb9-4275-44fb-b283-fa8dc6cfb928.png)

After:

![Screenshot 2023-02-16 at 16 52 22](https://user-images.githubusercontent.com/1121616/219433415-0c164edf-7eef-492f-95ce-9632ab180419.png)

And light mode:
![Screenshot 2023-02-16 at 16 53 28](https://user-images.githubusercontent.com/1121616/219433690-071b4c1b-3205-4e56-a81d-c00a8495f5bb.png)


It took a bit of doing, and could probably use some more tweaking, but I think it's good to have an up-to-date list of integrations and feature them a bit more prominently.

One thing to note is that I've changed from svg's that were built into the code to `<img>` tags. The new way will take a bit longer to load, but the files are very small, so I think it will be basically negligible, and be easier to maintain for us until we invest in some automation to build it into the static HTML again.