fix: Respect tables and skip_tables config in PostgreSQL source plugin (#8791)
Before this change, the PostgreSQL source plugin always synced all tables in the source database. This would lead to a panic if any table filters were applied, because the resources sent to the destination would be for tables that the destination does not have in its list of tables.

With this fix, users can choose which tables to sync using a source config, like usual. The default is still to sync all tables.