fix: Use the correct path for test file (#8939)
#### Summary

When using the S3 destination plugin, cloudquery failed with an Access
Denied error around the test file despite correct permissions.

Beforehand the code used the plugins path (i.e. cloudquery/s3). That
means it doesn't actually test the correct permissions. It should use
the output path that is provided by the user.

<!--
Use the following steps to ensure your PR is ready to be reviewed

- [ ] Read the [contribution guidelines](../blob/main/CONTRIBUTING.md)
🧑‍🎓
- [ ] Run `make lint` to ensure the proposed changes follow the coding
style 🚨 (install golangci-lint
[here](https://golangci-lint.run/usage/install/#local-installation))
- [ ] Run `make test` to ensure the proposed changes pass the tests 🧪
- [ ] If changing a source plugin run `make gen` to ensure docs are up
to date 📝
- [ ] Ensure the status checks below are successful ✅
--->