fix(postgresql): Normalize types in Postgres destination (#9143)
Fixes https://github.com/cloudquery/cloudquery/issues/9141. This will
allow CloudQuery to sync to tables that had their schemas created by
some external process, as long as the column can be normalized to a
CloudQuery type. This is mostly useful for Postgres -> Postgres sync use
cases right now. As we improve the internal type system, we should be
able to do away with these training wheels and create the same column
types on the destination as we find on the source for such cases.

The type normalization is copied from the [Postgres
source](https://github.com/cloudquery/cloudquery/blob/main/plugins/source/postgresql/client/types.go)