feat(azure): Discover resource groups and namespaces in parallel (#9382)
#### Summary

Related to https://github.com/cloudquery/cloudquery/issues/9269. I'm trying to reproduce https://github.com/cloudquery/cloudquery/issues/9269 with many subscriptions.
While doing so (I simulated a run on 100 subscriptions equivalent to our playground subscription) I noticed it takes a long time for the plugin to start syncing when I increase the subscriptions count.
The reason is that we discover resource groups and registered namespaces in serial.

This PR updates the code to do it in parallel. It improves the discovery time on 100 subscriptions from ~7m to ~30s.


<!--