fix(website): Default DESTINATION_NAME to postgresql (#9673)
In our example source configs, we currently use `postgres` as the example destination name. But in the postgresql destination, the name is `postgresql`. This can cause some confusion for users if they copy-paste from the examples, so it would be best to keep this consistent. This change makes it `postgresql` in all cases.