fix(policies-cloudtrail): Add region criteria to inner join (#10246)
#### Summary

This PR addresses inadequate criteria in the inner join of the policy
query that checks whether AWS CloudTrail is enabled in all regions,
[enabled_in_all_regions.sql](https://github.com/cloudquery/cloudquery/blob/main/plugins/source/aws/policies/queries/cloudtrail/enabled_in_all_regions.sql)

Specifically, the fix handles the situation of a single global
CloudTrail in a target AWS account by adding `region` criteria to the
`inner join`. It also corrects the results when an Organization-wide
CloudTrail is in place, by adding `account_id` criteria to the `inner
join`.

This bug is identified in #10245.