doc(deployment): Airflow deployment (#10370)
A basic guide on how to orchestrate CloudQuery syncs using Airflow and Kubernetes. There is a lot more that could be said here, but hopefully this can serve as a useful starting point to those wishing to use Airflow. In the future we can also follow up with more operators.