fix(aws): Explicitly set region in all SDK calls (#10453)
This updates all AWS resolvers to explicitly set the region in SDK
calls. This also now gets validated by our mock tests. We are doing this
so that in a follow-up release we can reduce memory usage for AWS
clients by ~20x by not needing to keep a separate client for every
region.

Part of #10452

---------

Co-authored-by: Herman Schaaf <hermanschaaf@gmail.com>
Co-authored-by: Kemal Hadimli <disq@users.noreply.github.com>