feat(cli): Move CLI to use plugin-pb instead of plugin-sdk (#10565)
Use new plugin-pb library instead of plugin-sdk, and move plugin-management related code from the SDK to here.

Testing needs to be improved but I hope to do so in the next versions of the protocol. 

Unrelated note - the more experienced we gather with source/dest plugin it becomes more clear that a single plugin might be a better abstraction then source/dest and will remove code duplication and enable better testing and devex (not something to do as we already have lot's of changes but to keep in mind).