doc(website): Add query examples to Azure documentation (#11347)
Similar to https://github.com/cloudquery/cloudquery/pull/11259 but for Azure.

The script had to be copied so that we don't create a dependency on another directory in the monorepo. It's either this, or we need to extract it to somewhere like the plugin-sdk, but I don't think we want this to be used as a library just yet.