test(duckdb): Add debug logs (#11660)
#### Summary

Related to https://github.com/cloudquery/cloudquery/issues/11550. Adds some debug logging to the DuckDB tests.
Hopefully we can trace back which query doesn't work. I initially thought the delete stale removes too much as the error is:
```
after overwrite expected 1 resource, got 0
```

but always got 1 row affected in the delete operation so 🤷  (the test is flaky on my machine too).
I would like to let the CI run a bit with the debug logs to see if it sheds more information on this issue.

<!--